chore: merge check and sync utils
