--- lib/i18n/check-utils.mjs
@@ -1,17 +1,0 @@
-import fs from 'fs'
-
-// get i18n config variables from package.json
-const { config } = JSON.parse(fs.readFileSync('package.json'))
-
-export const isTranslationTaskNeeded = () => {
-  const tempTranslations = JSON.parse(fs.readFileSync(`${config.i18n.locales_dir_path}/temp.json`))
-  const tempKeys = Object.keys(tempTranslations)
-
-  const sourceTranslations = JSON.parse(
-    fs.readFileSync(`${config.i18n.locales_dir_path}/${config.i18n.default_language_filename}.json`)
-  )
-
-  const idExists = id => Boolean(sourceTranslations[id])
-
-  return tempKeys.length > 0 && !tempKeys.every(idExists)
-}
