+++ lib/i18n/sync-utils.mjs
@@ -7,6 +7,19 @@ const execute = promisify(exec)
 // get i18n config variables from package.json
 const { config } = JSON.parse(readFileSync('package.json'))
 
+export const isTranslationTaskNeeded = () => {
+  const tempTranslations = JSON.parse(readFileSync(`${config.i18n.locales_dir_path}/temp.json`))
+  const tempKeys = Object.keys(tempTranslations)
+
+  const sourceTranslations = JSON.parse(
+    readFileSync(`${config.i18n.locales_dir_path}/${config.i18n.default_language_filename}.json`)
+  )
+
+  const idExists = id => Boolean(sourceTranslations[id])
+
+  return tempKeys.length > 0 && !tempKeys.every(idExists)
+}
+
 /**
  * Captures both formats: with app prefixes like "employer-branding-components-8WKgF6" and without like "ABC123"
  *
