+++ lib/i18n/sync.mjs
@@ -1,18 +1,18 @@
-import fs from 'fs'
+import { readdirSync, readFileSync, rename, writeFile } from 'node:fs'
 import path from 'path'
 import util from 'util'
 
 import { getNewToStaleIdMap } from './sync-utils.mjs'
 
 // get i18n config variables from package.json
-const { config } = JSON.parse(fs.readFileSync('package.json'))
+const { config } = JSON.parse(readFileSync('package.json'))
 
-const jsonsInDir = fs
-  .readdirSync(config.i18n.locales_dir_path)
-  .filter(file => path.extname(file) === '.json')
+const jsonsInDir = readdirSync(config.i18n.locales_dir_path).filter(
+  file => path.extname(file) === '.json'
+)
 
 const locales = jsonsInDir.reduce((acc, file) => {
-  const fileData = fs.readFileSync(path.join(config.i18n.locales_dir_path, file))
+  const fileData = readFileSync(path.join(config.i18n.locales_dir_path, file))
   const json = JSON.parse(fileData.toString())
   // don't use contextualized locales file when syncing
   if (file !== `contextualized-${config.i18n.default_language_filename}.json`) {
@@ -95,7 +95,7 @@ Object.entries(targetLocales).forEach(async ([localeName, localeData]) => {
     )
   }
 
-  fs.writeFile(
+  writeFile(
     `${config.i18n.locales_dir_path}/${localeName}.json`,
     JSON.stringify(newTarget, null, 2),
     'utf8',
@@ -114,7 +114,7 @@ Object.entries(targetLocales).forEach(async ([localeName, localeData]) => {
 })
 
 // rename temp file to source file name
-fs.rename(
+rename(
   `${config.i18n.locales_dir_path}/temp.json`,
   `${config.i18n.locales_dir_path}/${config.i18n.default_language_filename}.json`,
   () => {
